challenges/web_symple-unzipper/private/private/requirements.txt
fastapi~=0.78.0
patool~=1.12
python-multipart~=0.0.5
uvicorn~=0.17.6
